pyproject.toml
[tool.poetry]
name = "discordbotblaz"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.6.2"
async-timeout = "^3.0.1"
chardet = "^3.0.4"
colormath = "^3.0.0"
"discord.py" = "^1.4.1"
Flask = "^1.1.2"
idna = "^2.10"
numpy = "^1.19.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
